\name{BGSIMD-package}
\alias{BGSIMD-package}
\alias{BGSIMD}
\docType{package}
\title{
Efficient Block Gibbs Sampler with Data from an  Incomplete Multinomial Distribution }
\description{
Implement an efficient block Gibbs sampler for
Bayesian analysis with
incomplete data from a multinomial distribution taking values
from the k categories 1,2,...,k, where data
are assumed to miss at random and each missing datum
belongs to one and only one of m distinct non-empty proper
 subsets A1, A2,..., Am of
1,2,...,k and the k categories are labelled such that only
consecutive A's may overlap.
}
\details{
\tabular{ll}{
Package: \tab BGSIMD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-06\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Kwang Woo Ahn <kwooahn@mcw.edu> and Kung-Sik Chan <kung-sik-chan@uiowa.edu>
}
\references{
Ahn, K. W. and Chan, K. S. (2007)  Efficient Markov chain Monte Carlo with incomplete multinomial data, Technical report 382, The University of Iowa
}
\keyword{ package }
